# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zenku API
  version: 1.0.0
  description: |
    Schema-driven REST API for Zenku. All CRUD endpoints are auto-generated from the
    Prisma/ZenStack schema. Model names in paths use lowercase (e.g., `product`, `task`).
servers:
  - url: http://localhost:3001
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Attachment:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        storedName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        entityModel:
          type: string
          nullable: true
        entityId:
          type: string
          nullable: true
        url:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/{model}/batch:
    parameters:
      - name: model
        in: path
        required: true
        schema:
          type: string

    delete:
      summary: Batch delete records
      tags: [CRUD]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Records deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer

    patch:
      summary: Batch update records
      tags: [CRUD]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, data]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                data:
                  type: object
      responses:
        '200':
          description: Records updated
  /api/{model}/export:
    parameters:
      - name: model
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Export records as CSV
      tags: [CRUD]
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv]
            default: csv
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # ── File Upload ──────────────────────────────────────────────────────────────
  /api/upload:
    post:
      summary: Upload a file
      tags: [Files]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Max 10MB, any file type
                entityModel:
                  type: string
                  description: Model name to associate the file with
                entityId:
                  type: string
                  description: Record ID to associate the file with
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '413':
          description: File exceeds 10MB limit
  /api/uploads/{storedName}:
    get:
      summary: Serve an uploaded file
      tags: [Files]
      parameters:
        - name: storedName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
        '404':
          description: File not found

  # ── AI ───────────────────────────────────────────────────────────────────────
...